MAJOR MIGRATION - separated projects into acu and symp and shared folder
