+++ README.md
@@ -1,7 +1,6 @@
 # Predict Undesirable Cancer Events
 
-Build models to predict undesirable cancer events (i.e. symptom deterioration, emergency department visits, etc).
-Pipelines for data preparation, model development, and model evaluation. 
+Pipelines for data preparation, model development, and model evaluation for prediction of undesirable cancer events (i.e. symptom deterioration, emergency department visits, etc).
 
 # Data Location
 Please see [ml4oncology/make-clinical-dataset](https://github.com/ml4oncology/make-clinical-dataset).
@@ -9,8 +8,6 @@ Please see [ml4oncology/make-clinical-dataset](https://github.com/ml4oncology/ma
 # Getting started
 ```bash
 git clone https://github.com/ml4oncology/PredUCE
-conda env create -f envs/<env_name>.yaml
-conda activate aim2reduce
 pip install -e ".[dev]"
 
 # optional
