--- src/symp/models/LSTM.py
@@ -1,22 +1,0 @@
-import torch.nn as nn
-
-
-class LSTM(nn.Module):
-    def __init__(self, input_size, hidden_size, output_size, num_layers=1, dropout=0.2):
-        # super(LSTMModel, self).__init__()
-        super().__init__()
-        self.hidden_size = hidden_size
-        self.num_layers = num_layers
-        self.dropout = dropout
-
-        self.lstm = nn.LSTM(
-            input_size, hidden_size, num_layers, batch_first=True, dropout=dropout
-        )
-        self.outlayer = nn.Linear(hidden_size, output_size)
-        self.relu = nn.ReLU()
-        self.sigmoid = nn.Sigmoid()
-
-    def forward(self, full_df):
-        x, _ = self.lstm(full_df)
-        x = self.outlayer(x)
-        return x
