+++ symp/env.yaml
