+++ symp/scripts/R/bootstrap.R
