+++ symp/scripts/ed_association.py
