+++ symp/scripts/train_pipeline.py
