+++ symp/src/util.py
